# apps/hermes/swagger.yaml
swagger: "2.0"
info:
  title: Follow Email Backend API
  description: |
    A comprehensive email management and AI-powered analysis API that provides:
    - Email synchronization from multiple providers (Gmail, Outlook)
    - AI-powered email analysis and response generation
    - Queue management and background processing
    - Privacy compliance and GDPR features
  version: "1.0.0"
  contact:
    name: Follow Email Backend Team
    email: follow.email07@gmail.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
host: api.follow.email
basePath: /api/v1
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  bearer:
    type: apiKey
    name: Authorization
    in: header
    description: |
      Bearer token authentication. Use format: "Bearer <token>"
      Tokens are issued by Clerk authentication service.
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: |
      Bearer token authentication. Use format: "Bearer <token>"
      Tokens are issued by Clerk authentication service.
...
paths:
...
  /auth/login:
    post:
      tags:
        - auth
      summary: User login
      description: Authenticates user and returns JWT token
      parameters:
        - name: body
          in: body
          description: User login credentials
          required: true
          schema:
            type: object
            required:
              - email
              - password
            properties:
              email:
                type: string
                format: email
                example: "user@example.com"
              password:
                type: string
                example: "securePassword123"
      responses:
        200:
          description: Login successful
          schema:
            type: object
            properties:
              token:
                type: string
                example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
              user:
                type: object
                properties:
                  id:
                    type: string
                    example: "user_123456"
                  email:
                    type: string
                    format: email
                    example: "user@example.com"
                  first_name:
                    type: string
                    example: "John"
                  last_name:
                    type: string
                    example: "Doe"
        400:
          description: Bad request - validation errors
          schema:
            $ref: '#/definitions/ErrorResponse'
        401:
          description: Unauthorized - invalid credentials
          schema:
            $ref: '#/definitions/ErrorResponse'
        500:
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorResponse'
...
  /gmail/consent/callback:
    get:
      tags:
        - gmail
      summary: Gmail OAuth callback
      description: Handles the OAuth callback from Google (called automatically)
      parameters:
        - name: code
          in: query
          description: Authorization code from Google
          required: true
          type: string
        - name: state
          in: query
          description: State parameter for security
          required: true
          type: string
        - name: error
          in: query
          description: Error from Google if authorization failed
          required: false
          type: string
      responses:
        200:
          description: Gmail access granted successfully
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Gmail access granted successfully"
              user_email:
                type: string
                format: email
                example: "user@gmail.com"
              consent_date:
                type: string
                format: date-time
                example: "2024-01-15T10:30:00Z"
              sync_enabled:
                type: boolean
                example: true
        400:
          description: Bad request - invalid callback parameters or OAuth error
          schema:
            type: object
            properties:
              error:
                type: string
                example: "OAuth authorization failed"
              details:
                type: string
                example: "access_denied"
        500:
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorResponse'
...
definitions:
...
  ErrorResponse:
    type: object
    properties:
      error:
        type: string
        description: Error code or type
        example: "INVALID_REQUEST"
      message:
        type: string
        description: Human-readable error message
        example: "The request parameters are invalid"
      code:
        type: string
        description: Specific error code for programmatic handling
        example: "VALIDATION_ERROR"
      timestamp:
        type: string
        format: date-time
        description: Timestamp when the error occurred
        example: "2023-01-01T12:00:00Z"
      request_id:
        type: string
        description: Unique identifier for the request (for debugging)
        example: "req_550e8400-e29b-41d4-a716-446655440000"